Add action to generate certificates
Feature request to add an action to generate
certificates along with private key. Currently
the charms can relate to tls-certificates-operator
charm and request for certificates via the integration.
An action is required to request for certificate by
a non-charmed application.
Microstack (aka sunbeam) [1] is one such example which
requests certificate via actions.

[1] https://github.com/openstack-snaps/snap-microstack