build: unpin pytest-asyncio
This was originally pinned due to an incompatibility between pytest-operator and pytest-asyncio. pytest-operator now pins the version it requires, so we shouldn't.

By unpinning the version, we will be notified by renovate if pytest-operator ever supports a newer version of the library in the future, instead of holding it back.

Furthermore, we will stop getting Renovate PRs (which always fail) that are attempting to update pytest-asyncio when new versions of it are released.